Update cli to use conartist.
