--- packages/cli/README.md
@@ -1,37 +1,0 @@
-# SkateJS CLI
-
-> Hit the ground skating!
-
-## Installing
-
-```
-npm install @skatejs/cli
-```
-
-## Using
-
-You can run the CLI simply by running `skatejs`.
-
-```
-$ skatejs
-
-? What are you building? app
-? What renderer do you want to use? preact
-
-Generating package.json...
-Writing files...
-NPM installing...
-
-Done!
-```
-
-### Tests
-
-You get a working Jest setup out of the box! Just run `npm test` or `jest`.
-
-### Bundling
-
-App bundles are built with `parcel-bundler`.
-
-* To dev, run `npm start`.
-* To bundle, run `npm build`.
