--- packages/config-core/bin.js
@@ -1,6 +1,0 @@
-#! /usr/bin/env node
-
-require("conartist").bin({
-  ...require("./package.json"),
-  conartist: require(".")
-});
