--- packages/config-core/package.json
@@ -1,40 +1,0 @@
-{
-  "author": "You <at@you.com>",
-  "bugs": "https://github.com/skatejs/skatejs/issues",
-  "description": "Default generator for core SkateJS packages.",
-  "homepage": "https://skatejs.netlify.com",
-  "keywords": [
-    "web",
-    "components",
-    "custom",
-    "elements",
-    "shadow",
-    "dom"
-  ],
-  "license": "MIT",
-  "name": "@skatejs/config-core",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": "skatejs/skatejs",
-  "scripts": {
-    "test": "jest"
-  },
-  "browser": "module/index.js",
-  "dependencies": {
-    "conartist": "^0.28.0",
-    "package-info": "^3.0.2"
-  },
-  "devDependencies": {},
-  "esnext": "esnext/index.js",
-  "files": [
-    "esnext",
-    "main",
-    "module"
-  ],
-  "main": "main/index.js",
-  "module": "module/index.js",
-  "source": "src/index.ts",
-  "bin": "./bin.js",
-  "version": "0.0.0"
-}
